aider: Added text shadow to logo CSS.
# Aider chat conversation:

USER: Can you create an outline for the logo using CSS?

ASSISTANT: To create an outline for the logo using CSS, we need to modify `index.html`. Here's the plan:

1. Add a text-shadow property to the .logo class.

Here are the *SEARCH/REPLACE* blocks:

index.html
```html
<<<<<<< SEARCH
    .logo {
        text-align: center;
        font-size: 48px;
        font-family: 'Linux Libertine O', sans-serif;
    }
=======
    .logo {
        text-align: center;
        font-size: 48px;
        font-family: 'Linux Libertine O', sans-serif;
        text-shadow: -1px 0 black, 0 1px black, 1px 0 black, 0 -1px black;
    }
>>>>>>> REPLACE
```

Let me know if this is what you had in mind!