aider: Decreased the space between the contact text and the logo by adding a margin-bottom property to the .logo-text class.
# Aider chat conversation:

USER: Can you decrease the space between the contact text and the logo?

ASSISTANT: To decrease the space between the contact text and the logo, we need to modify `index.html`. Here's the plan:

1. Add a `margin-bottom` property to the `.logo-text` class to reduce the space between the logo text and the contact text.

Here are the *SEARCH/REPLACE* blocks:

index.html
```html
<<<<<<< SEARCH
    .logo-text {
        text-shadow: -1px 0 black, 0 1px black, 1px 0 black, 0 -1px black;
    }
=======
    .logo-text {
        text-shadow: -1px 0 black, 0 1px black, 1px 0 black, 0 -1px black;
        margin-bottom: 5px;
    }
>>>>>>> REPLACE
```

Let me know if this is what you had in mind!