aider: Decreased the margin-bottom of the logo text to 0px.
# Aider chat conversation:

USER: I didn't see any decrease between the space for the logo and the contact text for some reason

ASSISTANT: I apologize for the mistake. It seems that the previous change didn't have the desired effect. Let's try again.

To decrease the space between the logo and the contact text, we can modify the `margin-bottom` property of the `.logo-text` class. Here's the plan:

1. Update the `.logo-text` class to set the `margin-bottom` property to a smaller value, such as 0px.

Here are the *SEARCH/REPLACE* blocks:

index.html
```html
<<<<<<< SEARCH
    .logo-text {
        text-shadow: -1px 0 black, 0 1px black, 1px 0 black, 0 -1px black;
        margin-bottom: 5px;
    }
=======
    .logo-text {
        text-shadow: -1px 0 black, 0 1px black, 1px 0 black, 0 -1px black;
        margin-bottom: 0px;
    }
>>>>>>> REPLACE
```

Let me know if this change has the desired effect.